REFACTOR #12 extract Balloon from HomePage
